New time estimation algorithm
The time a measurement takes is now estimated in a new way, which works
for any arbitrary number of sweeped parameters.
The old function was renamed to estimate_time_old and is still available in
case the new one turns out to be buggy.

Furthermore, traverse() now automatically calls estimate_time() so that
you always get an estimate of how long your measurement will take, even
if you do not explicitly call estimate_time() in your measurement
script.
Finally, the comment/docstring at the top of the traverse() function has
been converted to proper docstring syntax.