Finished Keithley 236 plugin
The Keithley 236 plugin now works and has held up to preliminary
testing, although it is not thoroughly tested yet and does not yet check
for errors after every command which it probably should. Also some
features which will not be used but are really easily and quickly
implementable have been partially implementers (getters for parameters
that will not be used, but not the setters and parameter
specifications).