Fixed non-working compliance check for Keithley 236
The function get_hit_compliance() was wrongly implemented due to
misinterpretation of the documentation. In the previous implementation
it would return something that can be set with the M command and would
be always false for most common configurations. Now I have figured out
how to really check if the compliance has been hit by polling the serial
poll byte and implemented get_hit_compliance() accordingly. Note that it
does not indicate whether the instrument currently hits the compliance,
but rather whether the instrument has hit the compliance since the last
poll.