Fixed C-V measurement script
C-V measurement script now works! Also it can now be both executed as a
script (which automatically starts a measurement with default settings)
or imported as a module (which does not automatically start a
measurement but does make the measurement functions available). Also
there are two implemented ways of checking if a next voltage step should
be done: by checking if the compliance is hit or by checking if the
measured voltage is close enough to the source value.