Implemented error reporting routine
Implemented a routine that not only checks for errors and warnings, but
also reports them in a reasonably human-readable way. Errors are
raised while warnings are reported with the warnings module. The routine
has not yet been tested and is not yet automatically called after each
command, but this is planned after testing of the routine.