Refactored Keithley 236
Rearranged some methods in the Keithley 236 plugin so that all similar
functions (all setters, all getters etc.) are grouped together. Removed
the read() method, use get_value() instead. Added a non-optional "when"
argument to Keithley 236 exceptions to give a textual description of
what the code was trying to do when the exception occured.

Also fixed a wrongly named variable in the Bash script to start qtlab on
*nix.